pyproject.toml
[tool.poetry]
name = "django-sample"
version = "0.1.0"
description = ""
authors = ["dextto <dexter.haan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
celery = {extras = ["sqs"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
django = "^5.0.1"
mysqlclient = "^2.2.1"
djangorestframework = "^3.14.0"
pydantic = "^2.5.3"
django-celery-beat = "^2.7.0"
eventlet = "^0.38.0"
boto3 = "^1.35.75"
kombu = {extras = ["boto3"], version = "^5.4.2"}
pycurl = "^7.45.3"


[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.7"
black = "^23.12.1"
pytest = "^7.4.4"
pytest-dotenv = "^0.5.2"
mypy = "^1.8.0"
types-python-dateutil = "^2.8.19.14"
types-requests = "^2.31.0.20231231"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
